LayoutTests/http/tests/xmlhttprequest/origin-whitelisting-ip-addresses-with-subdomains-expected.txt

CONSOLE MESSAGE: line 16: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/get.txt. Origin http://localhost:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/get.txt. Origin http://localhost:8000 is not allowed by Access-Control-Allow-Origin.
Specifying that an IP address should match subdomains doesn't make sense. This test verifies that it doesn't do anything.

Testing: http://127.0.0.1:8000/xmlhttprequest/resources/get.txt (sync)
FAIL: Error: NetworkError: DOM Exception 19
Testing: http://127.0.0.1:8000/xmlhttprequest/resources/get.txt (async)
FAIL: 0
